massive documentation push and code reordering in files
